test: split operation-lease contracts under 300-line files
Generated-By: looper 0.11.0-beta.1 (runner=fixer, agent=grok-build)
